updateUser controller

+++ frontend/src/pages/UserListPage.js
@@ -22,7 +22,9 @@ const UserListPage = ({ history }) => {
   }, [dispatch, history, userInfo]);
 
   const deleteUserHandler = (id) => {
-    dispatch(deleteUser(id));
+    if (window.confirm('Are you sure?')) {
+      dispatch(deleteUser(id));
+    }
   };
 
   return (
@@ -66,19 +68,17 @@ const UserListPage = ({ history }) => {
                   }}
                 >
                   <LinkContainer to={`/user/${user._id}/edit`}>
-                    <>
-                      <Button variant='light' className='btn-sm'>
-                        <i className='fas fa-edit'></i>
-                      </Button>
-                      <Button
-                        variant='danger'
-                        className='btn-sm'
-                        onClick={() => deleteUserHandler(user._id)}
-                      >
-                        <i className='fas fa-trash'></i>
-                      </Button>
-                    </>
+                    <Button variant='light' className='btn-sm'>
+                      <i className='fas fa-edit'></i>
+                    </Button>
                   </LinkContainer>
+                  <Button
+                    variant='danger'
+                    className='btn-sm'
+                    onClick={() => deleteUserHandler(user._id)}
+                  >
+                    <i className='fas fa-trash'></i>
+                  </Button>
                 </td>
               </tr>
             ))}
